+++ gazebo/rendering/Scene.cc
@@ -56,6 +56,7 @@
 #include "gazebo/rendering/RFIDTagVisual.hh"
 #include "gazebo/rendering/VideoVisual.hh"
 #include "gazebo/rendering/TransmitterVisual.hh"
+#include "gazebo/rendering/SelectionObj.hh"
 
 #if OGRE_VERSION_MAJOR >= 1 && OGRE_VERSION_MINOR >= 8
 #include "gazebo/rendering/deferred_shading/SSAOLogic.hh"
@@ -731,8 +732,7 @@ UserCameraPtr Scene::GetUserCamera(const uint32_t index) const
 //////////////////////////////////////////////////
 void Scene::RemoveCamera(const std::string &_name)
 {
-  std::vector<CameraPtr>::iterator iter;
-  for (iter = this->dataPtr->cameras.begin();
+  for (auto iter = this->dataPtr->cameras.begin();
       iter != this->dataPtr->cameras.end(); ++iter)
   {
     if ((*iter)->Name() == _name)
@@ -3198,7 +3198,6 @@ void Scene::RemoveVisual(uint32_t _id)
       else
         ++piter;
     }
-
     this->RemoveVisualizations(vis);
 
     vis->Fini();
@@ -3355,8 +3354,13 @@ void Scene::RemoveVisualizations(rendering::VisualPtr _vis)
     rendering::VisualPtr childVis = _vis->GetChild(i);
     Visual::VisualType visType = childVis->GetType();
     if (visType == Visual::VT_PHYSICS || visType == Visual::VT_SENSOR
-       || visType == Visual::VT_GUI)
+        || visType == Visual::VT_GUI)
     {
+      // do not remove ModelManipulator's SelectionObj
+      // FIXME remove this hardcoded check, issue #1832
+      if (std::dynamic_pointer_cast<SelectionObj>(childVis) != NULL)
+        continue;
+
       toRemove.push_back(childVis);
     }
   }