+++ plugins/DepthCameraPlugin.cc
@@ -65,8 +65,9 @@ void DepthCameraPlugin::Load(sensors::SensorPtr _sensor,
         std::placeholders::_3, std::placeholders::_4, std::placeholders::_5));
 
   this->newImageFrameConnection = this->depthCamera->ConnectNewImageFrame(
-      boost::bind(&DepthCameraPlugin::OnNewImageFrame,
-        this, _1, _2, _3, _4, _5));
+      std::bind(&DepthCameraPlugin::OnNewImageFrame,
+        this, std::placeholders::_1, std::placeholders::_2,
+        std::placeholders::_3, std::placeholders::_4, std::placeholders::_5));
 
   this->parentSensor->SetActive(true);
 }