+++ Changelog.md
@@ -3,6 +3,9 @@
 1. Update depth camera sensor to publish depth values over a topic.
     * [Pull request #](https://bitbucket.org/osrf/gazebo/pull-request/)
     
+1. Add FollowerPlugin
+    * [Pull request #2085](https://bitbucket.org/osrf/gazebo/pull-request/2085)
+
 1. Fix circular dependency so that physics does not call the sensors API.
     * [Pull request #2089](https://bitbucket.org/osrf/gazebo/pull-request/2089)
     * [Issue #1516](https://bitbucket.org/osrf/gazebo/issues/1516)
@@ -40,6 +43,7 @@
     * [Pull request #2042](https://bitbucket.org/osrf/gazebo/pull-request/2042)
 
 1. Use more opaque pointers.
+    * [Pull request #2022](https://bitbucket.org/osrf/gazebo/pull-request/2022)
     * [Pull request #2025](https://bitbucket.org/osrf/gazebo/pull-request/2025)
     * [Pull request #2043](https://bitbucket.org/osrf/gazebo/pull-request/2043)
     * [Pull request #2044](https://bitbucket.org/osrf/gazebo/pull-request/2044)
@@ -124,6 +128,9 @@ flags from c functions. Windows did not like `CPPTYPE_*` in
 
 1. Light and light state for the server side
     * [Pull request #1920](https://bitbucket.org/osrf/gazebo/pull-request/1920)
+
+1. Add scale to model state so scaling works on log/playback.
+    * [Pull request #2020](https://bitbucket.org/osrf/gazebo/pull-request/2020)
 
 1. Added tests for WorldState
     * [Pull request #1968](https://bitbucket.org/osrf/gazebo/pull-request/1968)
@@ -308,6 +315,9 @@ using the same arguments used in the command line.
         * [Pull request #2059](https://bitbucket.org/osrf/gazebo/pull-request/2059)
         * [Pull request #2087](https://bitbucket.org/osrf/gazebo/pull-request/2087)
 
+    1. Support joint creation between links in nested model.
+        * [Pull request #2080](https://bitbucket.org/osrf/gazebo/pull-request/2080)
+
 1. Building editor updates
 
     1. Use opaque pointers in the building editor
@@ -334,6 +344,10 @@ using the same arguments used in the command line.
 ## Gazebo 6.0
 
 ### Gazebo 6.X.X (201X-XX-XX)
+
+1. Fix removing multiple camera sensors that have the same camera name.
+    * [Pull request #2081](https://bitbucket.org/osrf/gazebo/pull-request/2081)
+    * [Issue #1811](https://bitbucket.org/osrf/gazebo/issues/1811)
 
 1. Backport model editor toolbar fixed joint option from [pull request #1794](https://bitbucket.org/osrf/gazebo/pull-request/1794)
     * [Pull request #1957](https://bitbucket.org/osrf/gazebo/pull-request/1957)